requirements.txt
Pillow==12.0.0
qrcode==8.2
